add xmake debug session
